feat: implement app with smoke testing
Implementing an app that makes at least some sense. The app translates a
date expression in natural language into a python date and returns it.

Also as part of this commit I made the changes to the release process to implement some kind
of smoke testing and include them in the pipeline validation steps